renamed files, updated test script
